requirements.txt
python-pptx==0.6.21
requests==2.31.0
ollama==0.2.1
